# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Products & Prices Management API
  description: High-performance Hexagonal API for Product Management and Dynamic Multi-Currency Price Resolution.
  version: 1.0.0
  contact:
    name: Mango Engineering

servers:
  - url: http://localhost:8080
    description: Local development server

paths:
  /products:
    post:
      summary: Create a new product
      description: Registers a new product using a client-generated UUIDv7 identifier for CQRS/idempotency support.
      operationId: createProduct
      tags:
        - Products
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateProductRequest'
            example:
              id: "01952e42-7a57-7000-8000-000000000001"
              name: "Zapatillas deportivas"
              description: "Modelo 2025 edición limitada"
      responses:
        '201':
          description: Product successfully created.
          headers:
            Location:
              description: URI path to the created product resource.
              schema:
                type: string
                example: "/products/01952e42-7a57-7000-8000-000000000001"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateProductResponse'
              example:
                id: "01952e42-7a57-7000-8000-000000000001"
                name: "Zapatillas deportivas"
                description: "Modelo 2025 edición limitada"
        '400':
          description: Invalid request payload or domain invariant violation (e.g. non-UUIDv7 ID, blank name/description).
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
              example:
                type: "about:blank"
                title: "Invalid Request"
                status: 400
                detail: "Name cannot be blank"
                instance: "/products"

  /products/{id}/prices:
    get:
      summary: Query product prices
      description: Resolves either the active price on a specific date (when `date` is provided) or the chronological price history (when `date` is omitted), scoped by currency.
      operationId: getPrices
      tags:
        - Prices
      parameters:
        - name: id
          in: path
          required: true
          description: UUIDv7 identifier of the product.
          schema:
            type: string
            format: uuid
          example: "01952e42-7a57-7000-8000-000000000001"
        - name: date
          in: query
          required: false
          description: Optional query date in ISO-8601 format (YYYY-MM-DD). If provided, returns the single active price on that date. If omitted, returns the complete price history.
          schema:
            type: string
            format: date
          example: "2024-04-15"
        - name: currency
          in: query
          required: false
          description: ISO-4217 currency code (defaults to EUR if omitted).
          schema:
            type: string
            minLength: 3
            maxLength: 3
          example: "EUR"
      responses:
        '200':
          description: Price resolution successful (single active price object if date query param provided, or chronological array of price history items).
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/GetActivePriceResponse'
                  - type: array
                    items:
                      $ref: '#/components/schemas/PriceHistoryItemResponse'
              examples:
                activePrice:
                  summary: Single active price response (when date param provided)
                  value:
                    value: 99.99
                    currency: "EUR"
                priceHistory:
                  summary: Chronological price history response (when date param omitted)
                  value:
                    - id: "01952e42-7a57-7000-8000-000000000002"
                      value: 99.99
                      currency: "EUR"
                      initDate: "2024-01-01"
                      endDate: "2024-06-30"
                    - id: "01952e42-7a57-7000-8000-000000000003"
                      value: 149.99
                      currency: "EUR"
                      initDate: "2024-07-01"
                      endDate: "2024-12-31"
        '400':
          description: Invalid product ID, date format, or ISO currency code.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
              example:
                type: "about:blank"
                title: "Invalid Request"
                status: 400
                detail: "Id must be a valid UUIDv7"
                instance: "/products/01952e42-7a57-7000-8000-000000000001/prices"
        '404':
          description: No active price found for the specified product, date, and currency.

    post:
      summary: Add a price to a product
      description: Registers a discrete historical price interval for a product. Enforces non-overlapping validity ranges per product and currency.
      operationId: addPriceToProduct
      tags:
        - Prices
      parameters:
        - name: id
          in: path
          required: true
          description: UUIDv7 identifier of the product.
          schema:
            type: string
            format: uuid
          example: "01952e42-7a57-7000-8000-000000000001"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddPriceRequest'
            example:
              value: 99.99
              currency: "EUR"
              initDate: "2024-01-01"
              endDate: "2024-06-30"
      responses:
        '201':
          description: Price successfully added to the product.
          headers:
            Location:
              description: URI path to the created price resource.
              schema:
                type: string
                example: "/products/01952e42-7a57-7000-8000-000000000001/prices/01952e42-7a57-7000-8000-000000000002"
        '400':
          description: Invalid price payload, negative amount, invalid ISO currency, or initDate >= endDate.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
              example:
                type: "about:blank"
                title: "Invalid Request"
                status: 400
                detail: "Init date must be before end date"
                instance: "/products/01952e42-7a57-7000-8000-000000000001/prices"
        '409':
          description: Price validity range overlaps with an existing price for this product and currency.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
              example:
                type: "about:blank"
                title: "Price Overlap Conflict"
                status: 409
                detail: "Price validity period overlaps with an existing price for this product and currency"
                instance: "/products/01952e42-7a57-7000-8000-000000000001/prices"

  /products/{id}/prices/{priceId}:
    put:
      summary: Update a price
      description: Updates an existing price interval. Validates domain invariants and enforces temporal non-overlapping constraints per product and currency.
      operationId: updatePrice
      tags:
        - Prices
      parameters:
        - name: id
          in: path
          required: true
          description: UUIDv7 identifier of the product.
          schema:
            type: string
            format: uuid
          example: "01952e42-7a57-7000-8000-000000000001"
        - name: priceId
          in: path
          required: true
          description: UUIDv7 identifier of the price to update.
          schema:
            type: string
            format: uuid
          example: "01952e42-7a57-7000-8000-000000000002"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddPriceRequest'
            example:
              value: 149.99
              currency: "EUR"
              initDate: "2024-01-01"
              endDate: "2024-08-31"
      responses:
        '200':
          description: Price successfully updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PriceHistoryItemResponse'
              example:
                id: "01952e42-7a57-7000-8000-000000000002"
                value: 149.99
                currency: "EUR"
                initDate: "2024-01-01"
                endDate: "2024-08-31"
        '400':
          description: Invalid payload, negative amount, invalid ISO currency, or initDate >= endDate.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
        '404':
          description: Price or product not found.
        '409':
          description: Price validity range overlaps with another existing price for this product and currency.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'

    delete:
      summary: Delete a price
      description: Deletes an existing price record by its identifier.
      operationId: deletePrice
      tags:
        - Prices
      parameters:
        - name: id
          in: path
          required: true
          description: UUIDv7 identifier of the product.
          schema:
            type: string
            format: uuid
          example: "01952e42-7a57-7000-8000-000000000001"
        - name: priceId
          in: path
          required: true
          description: UUIDv7 identifier of the price to delete.
          schema:
            type: string
            format: uuid
          example: "01952e42-7a57-7000-8000-000000000002"
      responses:
        '204':
          description: Price successfully deleted.
        '400':
          description: Invalid UUID format for productId or priceId.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
        '404':
          description: Price not found for the specified product.

components:
  schemas:
    CreateProductRequest:
      type: object
      required:
        - name
        - description
      properties:
        id:
          type: string
          format: uuid
          description: Optional client-generated UUIDv7 (generated automatically by server if omitted).
          example: "01952e42-7a57-7000-8000-000000000001"
        name:
          type: string
          minLength: 1
          maxLength: 255
          description: Non-blank product name.
          example: "Zapatillas deportivas"
        description:
          type: string
          minLength: 1
          description: Non-blank product description.
          example: "Modelo 2025 edición limitada"

    CreateProductResponse:
      type: object
      required:
        - id
        - name
        - description
      properties:
        id:
          type: string
          format: uuid
          description: Unique product UUIDv7 identifier.
          example: "01952e42-7a57-7000-8000-000000000001"
        name:
          type: string
          description: Non-blank product name.
          example: "Zapatillas deportivas"
        description:
          type: string
          description: Non-blank product description.
          example: "Modelo 2025 edición limitada"

    AddPriceRequest:
      type: object
      required:
        - value
        - initDate
      properties:
        value:
          type: number
          format: decimal
          minimum: 0.01
          description: Positive price amount.
          example: 99.99
        currency:
          type: string
          minLength: 3
          maxLength: 3
          description: ISO-4217 currency code (defaults to EUR if omitted).
          example: "EUR"
        initDate:
          type: string
          format: date
          description: Start of validity period (inclusive).
          example: "2024-01-01"
        endDate:
          type: string
          format: date
          nullable: true
          description: End of validity period (inclusive), or null for open-ended pricing.
          example: "2024-06-30"

    GetActivePriceResponse:
      type: object
      required:
        - value
        - currency
      properties:
        value:
          type: number
          format: decimal
          description: Price amount.
          example: 99.99
        currency:
          type: string
          minLength: 3
          maxLength: 3
          description: ISO-4217 currency code.
          example: "EUR"

    PriceHistoryItemResponse:
      type: object
      required:
        - id
        - value
        - currency
        - initDate
      properties:
        id:
          type: string
          format: uuid
          description: Price identifier (UUIDv7).
          example: "01952e42-7a57-7000-8000-000000000002"
        value:
          type: number
          format: decimal
          description: Price amount.
          example: 99.99
        currency:
          type: string
          minLength: 3
          maxLength: 3
          description: ISO-4217 currency code.
          example: "EUR"
        initDate:
          type: string
          format: date
          description: Start of validity period (inclusive).
          example: "2024-01-01"
        endDate:
          type: string
          format: date
          nullable: true
          description: End of validity period (inclusive), or null for open-ended pricing.
          example: "2024-06-30"

    ProblemDetail:
      type: object
      required:
        - type
        - title
        - status
        - detail
      properties:
        type:
          type: string
          format: uri
          description: Problem type URI identifier.
          example: "about:blank"
        title:
          type: string
          description: Short human-readable summary of the problem.
          example: "Invalid Request"
        status:
          type: integer
          description: HTTP status code.
          example: 400
        detail:
          type: string
          description: Specific domain explanation of the error.
          example: "Name cannot be blank"
        instance:
          type: string
          format: uri-reference
          description: URI of the resource where the error occurred.
          example: "/products"